Support new Debian version of gmock
Debian moved gmock to /usr/src/googletest/googlemock, according to the
upstream move into the googletest project. This patch adds explicit
search patches for the old and new environment.

Also, it searches for gmock.cc instead of CMakeLists.txt, as the later
is too generic.

The GMOCK_INCLUDE_DIRS is needed for Ubuntu Trusty, because the gtest.h
in /usr/src/gmock/gtest/include/gtest/ (package google-mock) is
incompatible to the one in /usr/include/gtest/ (package libgtest-dev).