Wait for work queue to become empty in WaitForAllComputations
ComputeConstraintsForScan() is no longer the only thing that can
dispatch optimisation; now we have the lambdas from Add/Remove
CustomConstraints(). Thus, comparing the number of processed scans from
AddScan() with the one in the constraint builder is no longer a viable
way to determine if we are done or not. The work queue has to become
empty as well.