Implement fixed size transform interpolation buffer.
While using the TransformInterpolationBuffer in 3rd party code I noticed
that it seems to be incomplete. Neither does it have a buffer size
limit, nor does it "remove the oldest transform if the buffer size limit
is exceeded" as stated in the header.

The changes don't break compatibility with existing code:
unless explicitly set, the buffer size is unlimited.