the timing of the angular velocity update process is slow when there is no IMU data (#1648)
I think the timing of the angular velocity update process is slow when there is no IMU data.
Therefore, the posture used when calculating the amount of rotation by the ExtrapolateRotation
when estimating the posture of the PoseExtrapolator is the posture of the previous time and
the posture of the previous time, and there is one tempo delay.