Clean trimming of frozen trajectory data.
The deletion logic needs to take care of deleting all data that is
"exclusively" connected to the submaps that are to be removed. This is
achieved by looking at the data that is connected via constraints in
the graph.

Deleting a frozen trajectory (one without optimization constraints)
doesn't work that way and would leave dangling nodes in the graph.

This adds an additional logic that uses the `node_ids` field of the
submap instead of the constraints if the trajectory is frozen.

Signed-off-by: Michael Grupp <grupp@magazino.eu>