Drop superfluous ROS dependency on catkin (#1599)
This package's CMakeLists.txt doesn't reference catkin at all, and can
be regarded as a "pure CMake" package. In fact, the package.xml already
states that this package is built using cmake directly, but still
depends on catkin.

This change makes the package buildable without catkin, as would be
desired when building with colcon in ROS 2.